Rename `buildSchema` to `buildSchemaFromSDL` and export from root
